Rename v{2,3}filter to V{2,3}HTTPFilter to make it more clear what they are, and split them into their own files
